[C-3396] Refactor sign-up
