--- packages/web/src/pages/sign-up-page/SignUpPage.module.css
@@ -1,3 +1,0 @@
-.pageContent {
-  max-width: 1280px;
-}
