--- packages/web/src/pages/sign-up-page/pages/CreateEmailPage/CreateEmailPage.module.css
@@ -1,14 +1,0 @@
-.logo {
-  height: 100%;
-  width: 100%;
-  object-fit: contain;
-}
-.logo.desktop {
-  max-height: 160px;
-  max-width: 160px;
-}
-
-.logo.mobile {
-  max-height: 50px;
-  max-width: 180px;
-}
