--- packages/web/src/pages/sign-up-page/pages/CreateEmailPage/messages.tsx
@@ -1,17 +1,0 @@
-export const messages = {
-  title: 'Sign Up For Audius',
-  emailLabel: 'Email',
-  signUp: 'Sign Up Free',
-  haveAccount: 'Already have an account?',
-  signIn: 'Sign In',
-  subHeader: (
-    <>
-      Join the revolution in music streaming! <br /> Discover, connect, and
-      create on Audius.
-    </>
-  ),
-  socialsDividerText: 'Or, get started with one of your socials',
-  invalidEmail: 'Please enter a valid email.',
-  unknownError: 'Unknown error occurred.',
-  metaMaskNotRecommended: 'Signing up with MetaMask is not recommended.'
-}
